Refactor application handling: add job application retrieval by job ID, improve error handling, and update import paths.
